# backend/docs/openapi-ui-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "qswarm-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/test-design-runs/{run_id}": {
      "get": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Get Test Design Run",
        "operationId": "ui_get_test_design_run_api_v1_test_design_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UiTestDesignRunDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/analyze": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Analyze Requirements",
        "operationId": "ui_analyze_requirements_api_v1_test_design_runs__run_id__analyze_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "qswarm-web",
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/analysis": {
      "get": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Get Analysis",
        "operationId": "ui_get_analysis_api_v1_test_design_runs__run_id__analysis_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/prepare-plan": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Prepare Plan",
        "operationId": "ui_prepare_plan_api_v1_test_design_runs__run_id__prepare_plan_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "qswarm-web",
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/plan": {
      "get": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Get Plan",
        "operationId": "ui_get_plan_api_v1_test_design_runs__run_id__plan_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/approve-plan": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Approve Plan",
        "operationId": "ui_approve_plan_api_v1_test_design_runs__run_id__approve_plan_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "qswarm-web",
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/request-plan-revision": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Request Plan Revision",
        "operationId": "ui_request_plan_revision_api_v1_test_design_runs__run_id__request_plan_revision_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UiWorkspacePlanRevision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/generate-test-cases": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Generate Test Cases",
        "operationId": "ui_generate_test_cases_api_v1_test_design_runs__run_id__generate_test_cases_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "qswarm-web",
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/review-data": {
      "get": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Get Review Data",
        "operationId": "ui_get_review_data_api_v1_test_design_runs__run_id__review_data_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/request-revision": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Request Revision",
        "operationId": "ui_request_revision_api_v1_test_design_runs__run_id__request_revision_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UiWorkspaceRevision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/approve": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Approve Test Design",
        "operationId": "ui_approve_test_design_api_v1_test_design_runs__run_id__approve_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UiWorkspaceApprove"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/test-design-runs/{run_id}/publish": {
      "post": {
        "tags": [
          "ui-v1",
          "ui-v1-test-design-runs"
        ],
        "summary": "Ui Publish Test Design",
        "operationId": "ui_publish_test_design_api_v1_test_design_runs__run_id__publish_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "qswarm-web",
              "title": "Actor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UiArtifactVersionRef": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "artifactId": {
            "type": "string",
            "title": "Artifactid"
          },
          "content": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "planApproved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planapproved"
          },
          "planApprovedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planapprovedat"
          },
          "planApprovedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planapprovedby"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "artifactId",
          "content"
        ],
        "title": "UiArtifactVersionRef"
      },
...
      "UiSourceStoryRef": {
        "properties": {
          "storyKey": {
            "type": "string",
            "title": "Storykey"
          },
          "intakeArtifactId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intakeartifactid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "storyKey"
        ],
        "title": "UiSourceStoryRef"
      },
...
      "UiTestDesignRunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "storyKey": {
            "type": "string",
            "title": "Storykey"
          },
          "workflowName": {
            "type": "string",
            "title": "Workflowname"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "currentStep": {
            "type": "string",
            "title": "Currentstep"
          },
          "currentStage": {
            "type": "string",
            "title": "Currentstage"
          },
          "nextActions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nextactions"
          },
          "blockedReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockedreason"
          },
          "initiatedBy": {
            "type": "string",
            "title": "Initiatedby"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "sourceStory": {
            "$ref": "#/components/schemas/UiSourceStoryRef"
          },
          "requirementAnalysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UiArtifactVersionRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "testDesignPlan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UiArtifactVersionRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/UiTestDesignVersion"
            },
            "type": "array",
            "title": "Versions"
          },
          "reviewIssue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UiReviewIssueRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "testCaseRecords": {
            "items": {
              "$ref": "#/components/schemas/UiTestCaseRecord"
            },
            "type": "array",
            "title": "Testcaserecords"
          },
          "automationReadyTestCases": {
            "items": {
              "$ref": "#/components/schemas/UiTestCaseRecord"
            },
            "type": "array",
            "title": "Automationreadytestcases"
          },
          "approvalId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approvalid"
          },
          "productWorkspace": {
            "$ref": "#/components/schemas/UiProductWorkspace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "storyKey",
          "workflowName",
          "status",
          "currentStep",
          "currentStage",
          "nextActions",
          "initiatedBy",
          "sourceStory"
        ],
        "title": "UiTestDesignRunDetail",
        "description": "GET /api/v1/test-design-runs/{runId} \u2014 authoritative aggregate."
      },
...
      "UiWorkspaceApprove": {
        "properties": {
          "actorId": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Actorid",
            "default": "qswarm-web"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "UiWorkspaceApprove"
      },
      "UiWorkspacePlanRevision": {
        "properties": {
          "actorId": {
            "type": "string",
            "maxLength": 256,
            "title": "Actorid",
            "default": "qswarm-web"
          },
          "instruction": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Instruction"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "instruction"
        ],
        "title": "UiWorkspacePlanRevision"
      },
      "UiWorkspaceRevision": {
        "properties": {
          "actorId": {
            "type": "string",
            "maxLength": 256,
            "title": "Actorid",
            "default": "qswarm-web"
          },
          "instruction": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Instruction"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "default": "refine"
          }
        },
        "type": "object",
        "required": [
          "instruction"
        ],
        "title": "UiWorkspaceRevision"
      },
...
    }
  }
}